GHH: fixed a bug in bp, set layer_size_t to long

+++ tiny_cnn/util.h
@@ -71,7 +71,8 @@ namespace tiny_cnn {
 
 //typedef double float_t;
 typedef float float_t;
-typedef unsigned short layer_size_t;
+// typedef unsigned short layer_size_t;
+typedef unsigned long layer_size_t;
 typedef size_t label_t;
 typedef std::vector<float_t> vec_t;
